move integrity file
