--- packages/sdk/src/targets/index.ts
@@ -1,1 +1,0 @@
-export { checkIntegrity } from "./checkIntegrity"
