+++ packages/sdk/src/utils/targetIntegrity.ts
